+++ test/services/test_console.py
@@ -149,14 +149,6 @@ class ConsoleServiceTest(unittest.TestCase):
             "value": 25.0,
             "invoiceDesc": ""
         },
-        "member": {
-            "name": "",
-            "unit": "GB",
-            "currency": "USD",
-            "price": 5,
-            "value": 25.0,
-            "invoiceDesc": ""
-        },
         "realtime": {
             "name": "",
             "unit": "GB",
@@ -166,14 +158,6 @@ class ConsoleServiceTest(unittest.TestCase):
             "invoiceDesc": ""
         },
         "realtimeMessages": {
-            "name": "",
-            "unit": "GB",
-            "currency": "USD",
-            "price": 5,
-            "value": 25.0,
-            "invoiceDesc": ""
-        },
-        "realtimeBandwidth": {
             "name": "",
             "unit": "GB",
             "currency": "USD",
@@ -212,38 +196,9 @@ class ConsoleServiceTest(unittest.TestCase):
             "price": 5,
             "value": 25.0,
             "invoiceDesc": ""
-        },
-        "credits": {
-            "name": "",
-            "unit": "GB",
-            "currency": "USD",
-            "price": 5,
-            "value": 25.0,
-            "invoiceDesc": ""
         }
     },
-    "addons": {
-        "seats": {
-            "supported": True,
-            "planIncluded": 1.0,
-            "limit": 5.0,
-            "type": "numeric",
-            "currency": "USD",
-            "price": 5,
-            "value": 25.0,
-            "invoiceDesc": ""
-        },
-        "projects": {
-            "supported": True,
-            "planIncluded": 1.0,
-            "limit": 5.0,
-            "type": "numeric",
-            "currency": "USD",
-            "price": 5,
-            "value": 25.0,
-            "invoiceDesc": ""
-        }
-    },
+    "addons": {},
     "budgetCapEnabled": True,
     "customSmtp": True,
     "emailBranding": True,
@@ -383,14 +338,6 @@ class ConsoleServiceTest(unittest.TestCase):
                 "value": 25.0,
                 "invoiceDesc": ""
             },
-            "member": {
-                "name": "",
-                "unit": "GB",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            },
             "realtime": {
                 "name": "",
                 "unit": "GB",
@@ -400,14 +347,6 @@ class ConsoleServiceTest(unittest.TestCase):
                 "invoiceDesc": ""
             },
             "realtimeMessages": {
-                "name": "",
-                "unit": "GB",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            },
-            "realtimeBandwidth": {
                 "name": "",
                 "unit": "GB",
                 "currency": "USD",
@@ -446,38 +385,9 @@ class ConsoleServiceTest(unittest.TestCase):
                 "price": 5,
                 "value": 25.0,
                 "invoiceDesc": ""
-            },
-            "credits": {
-                "name": "",
-                "unit": "GB",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
             }
         },
-        "addons": {
-            "seats": {
-                "supported": True,
-                "planIncluded": 1.0,
-                "limit": 5.0,
-                "type": "numeric",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            },
-            "projects": {
-                "supported": True,
-                "planIncluded": 1.0,
-                "limit": 5.0,
-                "type": "numeric",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            }
-        },
+        "addons": {},
         "budgetCapEnabled": True,
         "customSmtp": True,
         "emailBranding": True,
@@ -510,22 +420,11 @@ class ConsoleServiceTest(unittest.TestCase):
     "billingStartDate": "2020-10-15T06:38:00.000+00:00",
     "billingCurrentInvoiceDate": "2020-10-15T06:38:00.000+00:00",
     "billingNextInvoiceDate": "2020-10-15T06:38:00.000+00:00",
-    "billingTrialStartDate": "2020-10-15T06:38:00.000+00:00",
     "billingTrialDays": 14.0,
     "billingAggregationId": "adbc3de4rddfsd",
     "billingInvoiceId": "adbc3de4rddfsd",
     "paymentMethodId": "adbc3de4rddfsd",
-    "billingAddressId": "adbc3de4rddfsd",
-    "backupPaymentMethodId": "adbc3de4rddfsd",
     "status": "active",
-    "remarks": "Pending initial payment",
-    "agreementBAA": "",
-    "programManagerName": "",
-    "programManagerCalendar": "",
-    "programDiscordChannelName": "",
-    "programDiscordChannelUrl": "",
-    "billingPlanDowngrade": "tier-1",
-    "billingTaxId": "",
     "markedForDeletion": True,
     "platform": "imagine",
     "projects": []