+++ test/services/test_organization.py
@@ -79,14 +79,6 @@ class OrganizationServiceTest(unittest.TestCase):
                 "value": 25.0,
                 "invoiceDesc": ""
             },
-            "member": {
-                "name": "",
-                "unit": "GB",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            },
             "realtime": {
                 "name": "",
                 "unit": "GB",
@@ -96,14 +88,6 @@ class OrganizationServiceTest(unittest.TestCase):
                 "invoiceDesc": ""
             },
             "realtimeMessages": {
-                "name": "",
-                "unit": "GB",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            },
-            "realtimeBandwidth": {
                 "name": "",
                 "unit": "GB",
                 "currency": "USD",
@@ -142,38 +126,9 @@ class OrganizationServiceTest(unittest.TestCase):
                 "price": 5,
                 "value": 25.0,
                 "invoiceDesc": ""
-            },
-            "credits": {
-                "name": "",
-                "unit": "GB",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
             }
         },
-        "addons": {
-            "seats": {
-                "supported": True,
-                "planIncluded": 1.0,
-                "limit": 5.0,
-                "type": "numeric",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            },
-            "projects": {
-                "supported": True,
-                "planIncluded": 1.0,
-                "limit": 5.0,
-                "type": "numeric",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            }
-        },
+        "addons": {},
         "budgetCapEnabled": True,
         "customSmtp": True,
         "emailBranding": True,
@@ -206,22 +161,11 @@ class OrganizationServiceTest(unittest.TestCase):
     "billingStartDate": "2020-10-15T06:38:00.000+00:00",
     "billingCurrentInvoiceDate": "2020-10-15T06:38:00.000+00:00",
     "billingNextInvoiceDate": "2020-10-15T06:38:00.000+00:00",
-    "billingTrialStartDate": "2020-10-15T06:38:00.000+00:00",
     "billingTrialDays": 14.0,
     "billingAggregationId": "adbc3de4rddfsd",
     "billingInvoiceId": "adbc3de4rddfsd",
     "paymentMethodId": "adbc3de4rddfsd",
-    "billingAddressId": "adbc3de4rddfsd",
-    "backupPaymentMethodId": "adbc3de4rddfsd",
     "status": "active",
-    "remarks": "Pending initial payment",
-    "agreementBAA": "",
-    "programManagerName": "",
-    "programManagerCalendar": "",
-    "programDiscordChannelName": "",
-    "programDiscordChannelUrl": "",
-    "billingPlanDowngrade": "tier-1",
-    "billingTaxId": "",
     "markedForDeletion": True,
     "platform": "imagine",
     "projects": []
@@ -300,14 +244,6 @@ class OrganizationServiceTest(unittest.TestCase):
                 "value": 25.0,
                 "invoiceDesc": ""
             },
-            "member": {
-                "name": "",
-                "unit": "GB",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            },
             "realtime": {
                 "name": "",
                 "unit": "GB",
@@ -317,14 +253,6 @@ class OrganizationServiceTest(unittest.TestCase):
                 "invoiceDesc": ""
             },
             "realtimeMessages": {
-                "name": "",
-                "unit": "GB",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            },
-            "realtimeBandwidth": {
                 "name": "",
                 "unit": "GB",
                 "currency": "USD",
@@ -363,38 +291,9 @@ class OrganizationServiceTest(unittest.TestCase):
                 "price": 5,
                 "value": 25.0,
                 "invoiceDesc": ""
-            },
-            "credits": {
-                "name": "",
-                "unit": "GB",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
             }
         },
-        "addons": {
-            "seats": {
-                "supported": True,
-                "planIncluded": 1.0,
-                "limit": 5.0,
-                "type": "numeric",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            },
-            "projects": {
-                "supported": True,
-                "planIncluded": 1.0,
-                "limit": 5.0,
-                "type": "numeric",
-                "currency": "USD",
-                "price": 5,
-                "value": 25.0,
-                "invoiceDesc": ""
-            }
-        },
+        "addons": {},
         "budgetCapEnabled": True,
         "customSmtp": True,
         "emailBranding": True,
@@ -427,22 +326,11 @@ class OrganizationServiceTest(unittest.TestCase):
     "billingStartDate": "2020-10-15T06:38:00.000+00:00",
     "billingCurrentInvoiceDate": "2020-10-15T06:38:00.000+00:00",
     "billingNextInvoiceDate": "2020-10-15T06:38:00.000+00:00",
-    "billingTrialStartDate": "2020-10-15T06:38:00.000+00:00",
     "billingTrialDays": 14.0,
     "billingAggregationId": "adbc3de4rddfsd",
     "billingInvoiceId": "adbc3de4rddfsd",
     "paymentMethodId": "adbc3de4rddfsd",
-    "billingAddressId": "adbc3de4rddfsd",
-    "backupPaymentMethodId": "adbc3de4rddfsd",
     "status": "active",
-    "remarks": "Pending initial payment",
-    "agreementBAA": "",
-    "programManagerName": "",
-    "programManagerCalendar": "",
-    "programDiscordChannelName": "",
-    "programDiscordChannelUrl": "",
-    "billingPlanDowngrade": "tier-1",
-    "billingTaxId": "",
     "markedForDeletion": True,
     "platform": "imagine",
     "projects": []